trustpilot load review button and code change

+++ app/code/local/Iksula/TrustPilotReviews/controllers/IndexController.php
@@ -22,25 +22,29 @@ class Iksula_TrustPilotReviews_IndexController extends Mage_Core_Controller_Fron
     }
 
     public function ajaxReviewAction(){
-     $nextpage_apiurl = $this->getRequest()->getPost('apiurl');
+      $nextpage_apiurl = $this->getRequest()->getPost('apiurl');
       if($nextpage_apiurl){
         Mage::getSingleton('core/session')->setNextPageApiUrl($nextpage_apiurl);
-        $block = Mage::app()->getLayout()->createBlock('trustpilotreviews/reviews')->getReviews();
-        foreach ($block['links'] as $links) {
+        $reviewsCollection = Mage::app()->getLayout()->createBlock('trustpilotreviews/reviews')->getReviews();
+        foreach ($reviewsCollection['links'] as $links) {
           if($links['rel'] == 'next-page'){
             $url = $links['href'];
           }
         }
-        $html = '';
-        foreach ($block['reviews'] as $reviews){
-          $html .=  $reviews['consumer']['displayName'].'<br>';
-          $html .=  '<image src="http://images-static.trustpilot.com/api/stars/'.$reviews['stars'].'/130x24.png" >'; 
-          $html .=  $reviews['title'].'<br>';
-          $html .=  $reviews['text'].'<br>';
-          $html .=  $reviews['companyReply']['text'].'<br>';
-          
+        if(array_key_exists('reviews',$reviewsCollection)){
+          $html = '';
+          foreach ($reviewsCollection['reviews'] as $reviews){
+            $html .=  $reviews['consumer']['displayName'].'<br>';
+            $html .=  '<image src="http://images-static.trustpilot.com/api/stars/'.$reviews['stars'].'/130x24.png" >'; 
+            $html .=  $reviews['title'].'<br>';
+            $html .=  $reviews['text'].'<br>';
+            $html .=  $reviews['companyReply']['text'].'<br>'; 
           }
-        echo json_encode(array('success'=>'true','apiurl'=>$url,'data'=>$html));
+          echo json_encode(array('success'=>'true','apiurl'=>$url,'data'=>$html));
+        }else{
+          echo json_encode(array('success'=>'false','message'=>'No More Reviews'));
+        }
+        
       }
     }
 }
